refactor(language-service): extract utility functions for reference and rename
This commit extracts utility functions and separates them from the core logic of the
references and rename builder.
